Create Covid 19 Model
The ongoing COVID-19 pandemic has resulted in the death of millions of people globally within two years. In an effort to better predict which U.S. states are more likely to suffer from high case rates in future pandemics, a logistic regression model was developed to determine which socio-economic factors have an influence on COVID-19 cases rates. The model was also developed to help predict which states would have low, medium, or high case rates. The regression model that was developed indicated that higher levels of population density, poverty levels, and percent of population with at least some college all contributed to a statistically significant increase in COVID-19 case rates. While, higher percentages of the population with the age of 65 or greater resulted in statistically significant lower case rates of COVID-19. The regression model was able to correctly predict the level of case rate (low, medium, or high) of each state, based on these socio-economic factors, with a 84% success rate using the training data set and with a 92% success rate in the validation data set.